ENYO-1654 :
##Issue
moon.ProgressBar needs a tooltip(popup) to show current progress.

##Fix
I extracted popup components and made PopupSupport.  Also, I organized slide and progressbar to apply this change.
WIth this change, app developer can use popup support in progresbar.
If user sets "popup:true", then popup will be shown.

Enyo-DCO-1.1-Signed-off-by: Suhyung Lee suhyung2.lee@lge.com
