+++ lib/ProgressBar/ProgressBar.js
@@ -10,6 +10,7 @@ var
 	Control = require('enyo/Control'),
 	Animator = require('enyo/Animator');
 
+var PopupSupport = require('moonstone/PopupSupport');
 /**
 * Fires when progress bar finishes animating to a position. No event-specific data
 * is sent with this event.
@@ -62,6 +63,11 @@ module.exports = kind(
 	* @private
 	*/
 	kind: Control,
+
+	/**
+	* @private
+	*/
+	mixins: [PopupSupport],
 
 	/**
 	* @private
@@ -126,7 +132,16 @@ module.exports = kind(
 		* @default 0
 		* @public
 		*/
-		bgProgress: 0
+		bgProgress: 0,
+
+		/**
+		* can make popup
+		*
+		* @type {Boolean}
+		* @default false
+		* @public
+		*/
+		popup: false
 	},
 
 	/**
@@ -192,6 +207,9 @@ module.exports = kind(
 		this.progress = this.clampValue(this.min, this.max, this.progress);
 		var p = this.calcPercent(this.progress);
 		this.updateBarPosition(p);
+		if(this.$.popup){
+			this.updatePopupPosition(p);
+		}
 	},
 
 	/**
@@ -258,5 +276,28 @@ module.exports = kind(
 	progressAnimatorComplete: function (inSender) {
 		this.doAnimateProgressFinish();
 		return true;
+	},
+
+	/**
+	* Initializes [progress]{@link module:moonstone/ProgressBar~ProgressBar#progress} at creation time.
+	*
+	* @private
+	*/
+	initProgress: function () {
+		this.updatePopupPosition(this.progress);
+	},
+
+	/**
+	* @private
+	*/
+	updatePopupPosition: function(val) {
+		var percent = this.calcPercent(val),
+			knobValue = (this.showPercentage && this.popupContent === null) ? percent : val;
+
+		this.$.popup.applyStyle('left', percent + '%');
+		this.$.popup.addRemoveClass('moon-slider-popup-flip-h', percent > 50);
+		this.$.popupLabel.addRemoveClass('moon-slider-popup-flip-h', percent > 50);
+
+		this.updatePopupLabel(knobValue);
 	}
 });
